Add i8, i16, i64, f32, char primitives

+++ examples/Float0/Float0.java
@@ -1,6 +1,6 @@
 public class Float0 {
-    public static void test(double i) {
-        double j = 10.0 / i;
-        assert Double.isFinite(j);
+    public static void test(float i) {
+        float j = 10.0f / i;
+        assert Float.isFinite(j);
     }
 }
